finish subnet view ; need to adjust poolid honeycomb-like layout
