--- src/components/IPmanagement/list/listmodule.module.scss
@@ -1,47 +1,0 @@
-.frame {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  height: 35px;
-  padding: 9px 0;
-  position: absolute;
-  left: 54px;
-  top: 3px;
-  width: 999px;
-}
-
-.text {
-  font-family: "Noto Sans", Helvetica, sans-serif;
-  font-size: 16px;
-  font-weight: 700;
-  letter-spacing: 1.6px;
-  width: 87px;
-  text-align: center;
-}
-
-.item {
-  font-family: "Noto Sans", Helvetica, sans-serif;
-  font-size: 16px;
-  font-weight: 400;
-  text-align: center;
-  height: 25px;
-  margin: -5px 0 -3px 0;
-  letter-spacing: 1.6px;
-}
-
-.label {
-  color: #2a6368;
-  font-weight: 700;
-  letter-spacing: 0.26px;
-}
-
-.value {
-  color: #000;
-  letter-spacing: 0.26px;
-}
-
-.divider {
-  width: 1px;
-  height: 30px;
-  background-color: #6bb7be;
-}
