+++ src/components/IPmanagement/poollist/index.ts
@@ -1,0 +1,1 @@
+export {PoolModule} from "./poollist";
