+++ src/components/IPmanagement/poollist/poollist.module.scss
@@ -1,0 +1,13 @@
+.poolcantainer {
+  margin-top: 10px;
+  padding: 12px;
+  background-color: #eefafc;
+  border-left: 4px solid #6bb7be;
+  transform: rotate(1deg) scale(1.02);
+  transition: all 0.3s ease;
+}
+
+.label {
+  font-weight: 700;
+  color: #2a6368;
+}
