+++ src/components/IPmanagement/subnet/SubnetModule.tsx
@@ -1,8 +1,9 @@
 import React from "react";
 import styles from "./subnet.module.scss";
-import { ListModule } from "../list";
+import { ListModule } from "../subnetlist";
 import { useGetSubnetsQuery } from "@/features/subnet/hooks/useSubnet";
 import { Subnet } from "@/features/subnet/types";
+import Button from "@/components/shared/Button";
 
 interface Props {
   className?: string;
@@ -10,26 +11,32 @@ interface Props {
 
 export const SubnetModule: React.FC<Props> = ({ className }) => {
   const { data: subnets, isLoading, isError, error } = useGetSubnetsQuery();
-
+  
   return (
     <div>
       {isLoading && <p>Loading...</p>}
       {isError && <p>Error: {(error as Error).message}</p>}
-      {subnets && subnets.length === 0 && <p>No subnets available</p>}
-      <div className={`${styles.component} ${className || ""}`}>
+      {/* {subnets && subnets.length === 0 && <p>No subnets available</p>} */}
+      
         {subnets &&
             subnets.map((subnet: Subnet, index: number) => (
-            <div key={index}>
-                <ListModule
-                    CIDR={subnet.cidr}
-                    NETMASK={subnet.netmask}
-                    GATEWAY={subnet.gateway}
-                    DC={subnet.id}
-                />
-            </div>
+              
+              <div className={`${styles.component} ${className || ""}`}>
+                <div key={index}>
+                  <div>
+                    <ListModule
+                        CIDR={subnet.cidr}
+                        NETMASK={subnet.netmask}
+                        GATEWAY={subnet.gateway}
+                        DC={subnet.id}
+                    />
+                  </div>
+                  <div style={{ marginBottom: "10px" }} />
+                </div>
+              </div>    
             ))}
-        </div>
-      <div className={styles.icon} />
+      
+      
     </div>
     
   );
