+++ src/components/IPmanagement/subnetlist/listmodule.module.scss
@@ -1,0 +1,74 @@
+.frame {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  height: 35px;
+  padding: 9px 0;
+  // position: absolute;
+  // left: 90px;
+  top: 3px;
+  width: 999px;
+  margin-left: 60px;
+}
+
+.text {
+  font-family: "Noto Sans", Helvetica, sans-serif;
+  font-size: 16px;
+  font-weight: 700;
+  letter-spacing: 1.6px;
+  width: 87px;
+  text-align: center;
+}
+
+.item {
+  font-family: "Noto Sans", Helvetica, sans-serif;
+  font-size: 16px;
+  font-weight: 400;
+  text-align: center;
+  height: 25px;
+  margin: -5px 0 -3px 0;
+  letter-spacing: 1.6px;
+}
+
+.label {
+  color: #2a6368;
+  font-weight: 700;
+  letter-spacing: 0.26px;
+}
+
+.value {
+  color: #000;
+  letter-spacing: 0.26px;
+}
+
+.divider {
+  width: 1px;
+  height: 30px;
+  background-color: #6bb7be;
+}
+
+.icon {
+  // background-image: url("https://c.animaapp.com/YZY4x0GN/img/icons-5@2x.png");
+  background-size: 100% 100%;
+  height: 26px;
+  width: 26px;
+  position: absolute;
+  top: 8px;
+  left: 14px;
+  border-radius: 0;
+  font-weight: 400;              // font-normal
+  font-size: 1rem;               // text-base
+  font-family: 'Noto_Sans', Helvetica;
+  letter-spacing: 2px;           // tracking-[2.00px]
+  
+  // transition: background-color 0.2s;
+}
+
+.honeycombWrapper {
+  margin-top: 10px;
+  padding: 12px;
+  background-color: #eefafc;
+  border-left: 4px solid #6bb7be;
+  transform: rotate(1deg) scale(1.02); /* Simulated honeycomb feel */
+  transition: all 0.3s ease;
+}
