+++ src/pages/IPmanagementPage/IPmanagementPage.module.scss
@@ -35,7 +35,7 @@
 
 .IPCard {
   width: 90%;
-  height: 80%;
+  height: 90%;
   margin: 0.25rem auto; // Increased top margin
   margin-left: 5%;
   margin-right: 5%;
